// src/resources/boxes.ts
export class Boxes extends APIResource {
  /**
   * @example
   * ```ts
   * const box = await client.boxes.create({
   *   type: 'linux',
   *   config: { env: { FOO: 'bar' } },
   * });
   * ```
   */
  create(body: BoxCreateParams, options?: RequestOptions): APIPromise<BoxCreateResponse> {
    return this._client.post('/boxes', { body, ...options });
  }

  /**
   * @example
   * ```ts
   * const box = await client.boxes.retrieve('id');
   * ```
   */
  retrieve(id: string, options?: RequestOptions): APIPromise<BoxRetrieveResponse> {
    return this._client.get(path`/boxes/${id}`, options);
  }

  /**
   * @example
   * ```ts
   * const boxes = await client.boxes.list();
   * ```
   */
  list(options?: RequestOptions): APIPromise<BoxListResponse> {
    return this._client.get('/boxes', options);
  }

  /**
   * @example
   * ```ts
   * const box = await client.boxes.delete('id');
   * ```
   */
  delete(id: string, options?: RequestOptions): APIPromise<BoxDeleteResponse> {
    return this._client.delete(path`/boxes/${id}`, options);
  }

  /**
   * @example
   * ```ts
   * const response = await client.boxes.createAndroid({
   *   config: { labels: { FOO: 'bar' }, envs: { FOO: 'bar' } },
   * });
   * ```
   */
  createAndroid(
    body: BoxCreateAndroidParams | null | undefined = {},
    options?: RequestOptions,
  ): APIPromise<BoxCreateAndroidResponse> {
    return this._client.post('/boxes/android', { body, ...options });
  }

  /**
   * @example
   * ```ts
   * const response = await client.boxes.createLinux({
   *   config: { envs: { FOO: 'bar' } },
   * });
   * ```
   */
  createLinux(
    body: BoxCreateLinuxParams | null | undefined = {},
    options?: RequestOptions,
  ): APIPromise<BoxCreateLinuxResponse> {
    return this._client.post('/boxes/linux', { body, ...options });
  }

  /**
   * @example
   * ```ts
   * const response = await client.boxes.start('id');
   * ```
   */
  start(id: string, options?: RequestOptions): APIPromise<BoxStartResponse> {
    return this._client.post(path`/boxes/${id}/start`, options);
  }

  /**
   * @example
   * ```ts
   * const response = await client.boxes.stop('id');
   * ```
   */
  stop(id: string, options?: RequestOptions): APIPromise<BoxStopResponse> {
    return this._client.post(path`/boxes/${id}/stop`, options);
  }
}

/**
 * A box instance that can be either Linux or Android type
 */
export type BoxCreateResponse = BoxCreateResponse.UnionMember0 | BoxCreateResponse.UnionMember1;
// ...
export type BoxRetrieveResponse = BoxRetrieveResponse.UnionMember0 | BoxRetrieveResponse.UnionMember1;

export namespace BoxRetrieveResponse {
  export interface UnionMember0 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for a Linux box instance
     */
    config: UnionMember0.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Linux
     */
    type: 'linux';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }

  export namespace UnionMember0 {
    /**
     * Configuration for a Linux box instance
     */
    export interface Config {
      /**
       * Browser configuration
       */
      browser: Config.Browser;

      /**
       * CPU cores allocated to the box
       */
      cpu: number;

      /**
       * Environment variables for the box
       */
      envs: Record<string, string>;

      /**
       * Key-value pairs of labels for the box
       */
      labels: Record<string, string>;

      /**
       * Memory allocated to the box in MB
       */
      memory: number;

      /**
       * Operating system configuration
       */
      os: Config.Os;

      /**
       * Storage allocated to the box in GB
       */
      storage: number;

      /**
       * Working directory path for the box
       */
      workingDir: string;
    }

    export namespace Config {
      /**
       * Browser configuration
       */
      export interface Browser {
        /**
         * Supported browser types for Linux boxes
         */
        type: 'chromium' | 'firefox' | 'webkit';

        /**
         * Browser version string (e.g. '12')
         */
        version: string;
      }

      /**
       * Operating system configuration
       */
      export interface Os {
        /**
         * Supported operating system types for Linux boxes
         */
        type: 'linux' | 'windows' | 'macos';

        /**
         * OS version string (e.g. 'ubuntu-20.04')
         */
        version: string;
      }
    }
  }

  export interface UnionMember1 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for an Android box instance
     */
    config: UnionMember1.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Android
     */
    type: 'android';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }

  export namespace UnionMember1 {
    /**
     * Configuration for an Android box instance
     */
    export interface Config {
      /**
       * Browser configuration
       */
      browser: Config.Browser;

      /**
       * CPU cores allocated to the box
       */
      cpu: number;

      /**
       * Environment variables for the box
       */
      envs: Record<string, string>;

      /**
       * Key-value pairs of labels for the box
       */
      labels: Record<string, string>;

      /**
       * Memory allocated to the box in MB
       */
      memory: number;

      /**
       * Android operating system configuration
       */
      os: Config.Os;

      /**
       * Storage allocated to the box in GB
       */
      storage: number;

      /**
       * Working directory path for the box
       */
      workingDir: string;
    }

    export namespace Config {
      /**
       * Browser configuration
       */
      export interface Browser {
        /**
         * Supported browser types for Android boxes
         */
        type: 'Chrome for Android' | 'UC Browser for Android';

        /**
         * Browser version string (e.g. '136')
         */
        version: string;
      }

      /**
       * Android operating system configuration
       */
      export interface Os {
        /**
         * Android operating system type
         */
        type: 'android';

        /**
         * Supported Android versions
         */
        version: '12' | '13';
      }
    }
  }
}

export type BoxListResponse = Array<BoxListResponse.UnionMember0 | BoxListResponse.UnionMember1>;

export namespace BoxListResponse {
  export interface UnionMember0 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for a Linux box instance
     */
    config: UnionMember0.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Linux
     */
    type: 'linux';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }

  export namespace UnionMember0 {
    /**
     * Configuration for a Linux box instance
     */
    export interface Config {
      /**
       * Browser configuration
       */
      browser: Config.Browser;

      /**
       * CPU cores allocated to the box
       */
      cpu: number;

      /**
       * Environment variables for the box
       */
      envs: Record<string, string>;

      /**
       * Key-value pairs of labels for the box
       */
      labels: Record<string, string>;

      /**
       * Memory allocated to the box in MB
       */
      memory: number;

      /**
       * Operating system configuration
       */
      os: Config.Os;

      /**
       * Storage allocated to the box in GB
       */
      storage: number;

      /**
       * Working directory path for the box
       */
      workingDir: string;
    }

    export namespace Config {
      /**
       * Browser configuration
       */
      export interface Browser {
        /**
         * Supported browser types for Linux boxes
         */
        type: 'chromium' | 'firefox' | 'webkit';

        /**
         * Browser version string (e.g. '12')
         */
        version: string;
      }

      /**
       * Operating system configuration
       */
      export interface Os {
        /**
         * Supported operating system types for Linux boxes
         */
        type: 'linux' | 'windows' | 'macos';

        /**
         * OS version string (e.g. 'ubuntu-20.04')
         */
        version: string;
      }
    }
  }

  export interface UnionMember1 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for an Android box instance
     */
    config: UnionMember1.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Android
     */
    type: 'android';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }

  export namespace UnionMember1 {
    /**
     * Configuration for an Android box instance
     */
    export interface Config {
      /**
       * Browser configuration
       */
      browser: Config.Browser;

      /**
       * CPU cores allocated to the box
       */
      cpu: number;

      /**
       * Environment variables for the box
       */
      envs: Record<string, string>;

      /**
       * Key-value pairs of labels for the box
       */
      labels: Record<string, string>;

      /**
       * Memory allocated to the box in MB
       */
      memory: number;

      /**
       * Android operating system configuration
       */
      os: Config.Os;

      /**
       * Storage allocated to the box in GB
       */
      storage: number;

      /**
       * Working directory path for the box
       */
      workingDir: string;
    }

    export namespace Config {
      /**
       * Browser configuration
       */
      export interface Browser {
        /**
         * Supported browser types for Android boxes
         */
        type: 'Chrome for Android' | 'UC Browser for Android';

        /**
         * Browser version string (e.g. '136')
         */
        version: string;
      }

      /**
       * Android operating system configuration
       */
      export interface Os {
        /**
         * Android operating system type
         */
        type: 'android';

        /**
         * Supported Android versions
         */
        version: '12' | '13';
      }
    }
  }
}
// ...
export type BoxCreateParams = BoxCreateParams.Variant0 | BoxCreateParams.Variant1;
// ...
export declare namespace Boxes {
  export {
    type BoxCreateResponse as BoxCreateResponse,
    type BoxRetrieveResponse as BoxRetrieveResponse,
    type BoxListResponse as BoxListResponse,
    type BoxDeleteResponse as BoxDeleteResponse,
    type BoxCreateAndroidResponse as BoxCreateAndroidResponse,
    type BoxCreateLinuxResponse as BoxCreateLinuxResponse,
    type BoxStartResponse as BoxStartResponse,
    type BoxStopResponse as BoxStopResponse,
    type BoxCreateParams as BoxCreateParams,
    type BoxCreateAndroidParams as BoxCreateAndroidParams,
    type BoxCreateLinuxParams as BoxCreateLinuxParams,
  };
}